# specs/001-todo-web-app/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Todo Full-Stack Web Application API
  description: Secure, multi-user todo application with JWT authentication
  version: 1.0.0
  contact:
    name: Development Team
    email: team@example.com
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.todoapp.com
    description: Production server
paths:
  /auth/register:
    post:
      summary: Register a new user
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                password:
                  type: string
                  format: password
                  minLength: 8
                  example: securePassword123
                firstName:
                  type: string
                  example: John
                lastName:
                  type: string
                  example: Doe
              required:
                - email
                - password
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                    example: 123e4567-e89b-12d3-a456-426614174000
                  email:
                    type: string
                    format: email
                    example: user@example.com
                  firstName:
                    type: string
                    example: John
                  lastName:
                    type: string
                    example: Doe
                  createdAt:
                    type: string
                    format: date-time
                    example: "2026-02-06T10:00:00Z"
        '400':
          description: Invalid input or user already exists
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
                    example: "Email already registered"
  /auth/login:
    post:
      summary: Login user and return JWT
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                password:
                  type: string
                  format: password
                  example: securePassword123
              required:
                - email
                - password
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
                    format: jwt
                    example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                  refreshToken:
                    type: string
                    format: jwt
                    example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                  expiresIn:
                    type: integer
                    example: 3600
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
                    example: "Invalid email or password"
  /auth/me:
    get:
      summary: Get current user information
      tags:
        - Authentication
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Current user information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /tasks:
    get:
      summary: Get all tasks for the authenticated user
      tags:
        - Tasks
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [pending, completed, all]
            default: all
          description: Filter tasks by status
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 10
            maximum: 100
          description: Maximum number of tasks to return
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            default: 0
          description: Number of tasks to skip
      responses:
        '200':
          description: List of tasks for the user
          content:
            application/json:
              schema:
                type: object
                properties:
                  tasks:
                    type: array
                    items:
                      $ref: '#/components/schemas/Task'
                  total:
                    type: integer
                    example: 5
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    post:
      summary: Create a new task for the authenticated user
      tags:
        - Tasks
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  maxLength: 200
                  example: "Buy groceries"
                description:
                  type: string
                  maxLength: 1000
                  example: "Milk, eggs, bread"
              required:
                - title
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
                    example: "Title is required"
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /tasks/{id}:
    get:
      summary: Get a specific task by ID
      tags:
        - Tasks
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Task ID
      responses:
        '200':
          description: Task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
                    example: "Task not found"
    put:
      summary: Update a task
      tags:
        - Tasks
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Task ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  maxLength: 200
                  example: "Updated task title"
                description:
                  type: string
                  maxLength: 1000
                  example: "Updated task description"
                status:
                  type: string
                  enum: [pending, completed]
                  example: "completed"
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
                    example: "Invalid status value"
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
    delete:
      summary: Delete a task
      tags:
        - Tasks
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Task ID
      responses:
        '204':
          description: Task deleted successfully
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /tasks/{id}/toggle:
    patch:
      summary: Toggle task completion status
      tags:
        - Tasks
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Task ID
      responses:
        '200':
          description: Task status toggled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174000
        email:
          type: string
          format: email
          example: user@example.com
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: Doe
        createdAt:
          type: string
          format: date-time
          example: "2026-02-06T10:00:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2026-02-06T10:00:00Z"
      required:
        - id
        - email
        - createdAt
        - updatedAt
    Task:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174000
        title:
          type: string
          maxLength: 200
          example: "Buy groceries"
        description:
          type: string
          maxLength: 1000
          example: "Milk, eggs, bread"
        status:
          type: string
          enum: [pending, completed]
          example: "pending"
        userId:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174001
        createdAt:
          type: string
          format: date-time
          example: "2026-02-06T10:00:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2026-02-06T10:00:00Z"
        completedAt:
          type: string
          format: date-time
          nullable: true
          example: "2026-02-06T11:00:00Z"
      required:
        - id
        - title
        - status
        - userId
        - createdAt
        - updatedAt
  responses:
    UnauthorizedError:
      description: Authentication required
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: string
                example: "Not authenticated"
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: string
                example: "Resource not found"